Extract generateJwt from archiving into its own module
