+++ lib/archiving.js
@@ -5,7 +5,7 @@ var request = require('request');
 var errors = require('./errors');
 var pkg = require('../package.json');
 var _ = require('lodash');
-var jwt = require('jsonwebtoken');
+var generateJwt = require('./generateJwt.js');
 
 // functions
 var api;
@@ -155,22 +155,10 @@ function Archive(config, properties) {
   };
 }
 
-var _generateJwt = function (config) {
-  var currentTime = Math.floor(new Date() / 1000);
-  var token = jwt.sign({
-    iss: config.apiKey,
-    ist: 'project',
-    iat: currentTime,
-    exp: currentTime + config.auth.expire
-  }, config.apiSecret);
-
-  return token;
-};
-
 var _generateHeaders = function (config) {
   return {
     'User-Agent': 'OpenTok-Node-SDK/' + pkg.version,
-    'X-OPENTOK-AUTH': _generateJwt(config),
+    'X-OPENTOK-AUTH': generateJwt(config),
     Accept: 'application/json'
   };
 };
